+++ applications/search-api/src/main/java/no/fdk/searchapi/controller/CatalogController.java
@@ -1,9 +1,9 @@
-package no.dcat.portal.query.controller;
+package no.fdk.searchapi.controller;
 
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import no.dcat.portal.query.ModelFormatter;
-import no.dcat.portal.query.SupportedFormat;
+import no.fdk.searchapi.ModelFormatter;
+import no.fdk.searchapi.SupportedFormat;
 import no.dcat.shared.Catalog;
 import no.dcat.shared.Dataset;
 import org.apache.jena.query.*;
